encapsulate panel tabs
